Ubrali if, nashli oshibku formata

+++ test/testkv.c
@@ -51,7 +51,7 @@ CTEST(solution_of_quadratic_equations, the_coefficients_are_not_correct)
     const int b = 1;
     const int c = 9;
     
-    double k1, k2;
+    float k1, k2;
     const int result = testkv(a, b, c, &k1, &k2);
    
     ASSERT_EQUAL(INVALID_ARGUMENTS, result);
